Fixes controllers and services

+++ src/main/java/com/simpleworkoutservice/simpleworkoutservice/dao/LiftSetRepository.java
@@ -1,9 +1,8 @@
 package com.simpleworkoutservice.simpleworkoutservice.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.simpleworkoutservice.simpleworkoutservice.entity.LiftSet;
 
-public interface LiftSetRepository extends JpaRepository<LiftSet, Integer> {
+public interface LiftSetRepository extends JpaRepository<LiftSet, Long> {
 
 }
