+++ src/main/java/com/simpleworkoutservice/simpleworkoutservice/dao/WorkoutRepository.java
@@ -1,13 +1,8 @@
 package com.simpleworkoutservice.simpleworkoutservice.dao;
 
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.simpleworkoutservice.simpleworkoutservice.entity.Exercise;
 import com.simpleworkoutservice.simpleworkoutservice.entity.Workout;
 
-public interface WorkoutRepository extends JpaRepository<Workout, Integer> {
-
-    // List<Workout> findAllByWeekId(int weekId);
+public interface WorkoutRepository extends JpaRepository<Workout, Long> {
 
 }
