+++ src/main/java/com/simpleworkoutservice/simpleworkoutservice/service/ExerciseService/ExerciseService.java
@@ -1,12 +1,8 @@
 package com.simpleworkoutservice.simpleworkoutservice.service.ExerciseService;
 
-import java.util.List;
-
 import com.simpleworkoutservice.simpleworkoutservice.entity.Exercise;
 import com.simpleworkoutservice.simpleworkoutservice.service.BaseService;
 
-public interface ExerciseService extends BaseService<Exercise, Integer> {
-
-    // List<Exercise> findAllByWorkoutId(int id);
+public interface ExerciseService extends BaseService<Exercise, Long> {
 
 }
