+++ src/main/java/com/simpleworkoutservice/simpleworkoutservice/service/WeekService/WeekService.java
@@ -1,13 +1,8 @@
 package com.simpleworkoutservice.simpleworkoutservice.service.WeekService;
 
-import java.util.List;
 import com.simpleworkoutservice.simpleworkoutservice.entity.Week;
 import com.simpleworkoutservice.simpleworkoutservice.service.BaseService;
 
-public interface WeekService extends BaseService<Week, Integer> {
-
-    // public List<Week> findAllByUserId(int id);
-
-    // public List<Week> findAllByProgramId(int programId);
+public interface WeekService extends BaseService<Week, Long> {
 
 }
